Move crypto functions to its own file
